# docs/assets/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/holding": {
      "get": {
        "description": "Retrieves all holdings for the authenticated user within a specified account.",
        "operationId": "HoldingController_getHoldings",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "description": "The ID of the account to retrieve holdings for.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holdings found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Holding"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get holdings for a specific account.",
        "tags": ["Holding"]
      }
    },
    "/holding/history": {
      "get": {
        "description": "Retrieves holding history for the given account. This is useful for displaying the holdings value over time.",
        "operationId": "HoldingController_getHoldingHistory",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "description": "The ID of the account to retrieve holding history for.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holding history found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityHistory"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get holding history for a specific account.",
        "tags": ["Holding"]
      }
    },
    "/holding/history/{id}": {
      "get": {
        "description": "Retrieves the value and performance history for a single holding given by its ID.",
        "operationId": "HoldingController_getSpecificHoldingHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the specific holding.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holding history found successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityHistory"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get history for a specific holding.",
        "tags": ["Holding"]
      }
    },
...
    "/holding/timeline/major": {
      "get": {
        "description": "Retrieves daily data points over the last 7 days for the major market indicies to support multi-line comparative charting templates.",
        "operationId": "HoldingController_getMajorIndicesTimeline",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully fetched major index timeline records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MajorIndexTimelineDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get 7-day historical performance for major market indices.",
        "tags": ["Holding"]
      }
    },
    "/holding/timeline/{id}": {
      "get": {
        "description": "Retrieves the value change over time of a holding given by it's ID.",
        "operationId": "HoldingController_getHoldingTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holding over time successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoricalDataPoint"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get timeline of a specific holding.",
        "tags": ["Holding"]
      }
    },
...
  },
  "info": {
    "title": "Sprout API",
    "description": "Welcome to the Sprout API documentation. This document provides a comprehensive guide to all available endpoints. Use the sections below to explore different parts of the API.\n\n  **Note:** If you wish to test the functionality, you must use a login endpoint to properly attach cookies to your browser.",
    "version": "v0.5.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Holding": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "accountId": {
            "type": "string",
            "description": "The Id of the account related to this holding."
          },
          "purchasePrice": {
            "type": "number",
            "description": "The numeric value converted to the user's preferred currency format. This overrides the original purchasePrice property.",
            "nullable": false
          },
          "costBasis": {
            "type": "number",
            "nullable": false,
            "description": "The numeric value converted to the user's preferred currency format. This overrides the original costBasis property."
          },
          "marketValue": {
            "type": "number",
            "description": "The numeric value converted to the user's preferred currency format. This overrides the original marketValue property.",
            "nullable": false
          },
          "description": {
            "type": "string",
            "description": "A description of what this holding is"
          },
          "shares": {
            "type": "number",
            "description": "Total number of shares, including fractional"
          },
          "symbol": {
            "type": "string",
            "description": "The symbol for this holding"
          }
        },
        "required": ["id", "accountId", "purchasePrice", "costBasis", "marketValue", "description", "shares", "symbol"]
      },
      "EntityHistoryDataPoint": {
        "type": "object",
        "properties": {
          "start": {
            "format": "date-time",
            "type": "string",
            "description": "When our data starts",
            "example": "2025-07-21T17:32:28Z"
          },
          "percentChange": {
            "type": "number"
          },
          "valueChange": {
            "type": "number"
          }
        },
        "required": ["start", "valueChange"]
      },
      "EntityHistory": {
        "type": "object",
        "properties": {
          "last1Day": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "last7Days": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "lastMonth": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "lastThreeMonths": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "lastSixMonths": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "lastYear": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "allTime": {
            "$ref": "#/components/schemas/EntityHistoryDataPoint"
          },
          "connectedId": {
            "type": "string",
            "description": "Some entity history data may have a connected Id of what it relates to. This could be something like an account Id."
          }
        },
        "required": ["last1Day", "last7Days", "lastMonth", "lastThreeMonths", "lastSixMonths", "lastYear", "allTime"]
      },
...
      "MajorIndexTimelinePoint": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-29T13:30:00.000Z"
          },
          "value": {
            "type": "number",
            "description": "The raw closing nominal price.",
            "example": 7580.06
          },
          "changePercent": {
            "type": "number",
            "description": "The percentage change relative to day one of the lookback window.",
            "example": 0.81
          }
        },
        "required": ["date", "value", "changePercent"]
      },
      "MajorIndexTimelineDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "description": "The official Yahoo Finance ticker index symbol code.",
            "example": "^GSPC"
          },
          "name": {
            "type": "string",
            "description": "The readable corporate clean name matching the index symbol.",
            "example": "S&P 500"
          },
          "color": {
            "type": "string",
            "description": "The hexadecimal color token string representing this brand asset.",
            "example": "#2196F3"
          },
          "timeline": {
            "description": "The historical performance point ledger tracking both raw price and performance percentages.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MajorIndexTimelinePoint"
            }
          }
        },
        "required": ["symbol", "name", "color", "timeline"]
      },
      "HistoricalDataPoint": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "The date of the record"
          },
          "value": {
            "type": "number",
            "description": "The numerical currency value"
          }
        },
        "required": ["date", "value"]
      },
...
    }
  }
}